backend/pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Post-discharge sepsis/deterioration catch via AI voice check-ins"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "motor>=3.5",
  "pymongo>=4.8",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "apscheduler>=3.10",
  "twilio>=9.0",
  "elevenlabs>=1.5",
  "websockets>=12.0",
  "google-generativeai>=0.8",
  "opensmile>=2.5",
  "numpy>=1.26",
  "soundfile>=0.12",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "audioop-lts>=0.2; python_version>='3.13'",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-mock>=3.14",
  "mongomock-motor>=0.0.29",
  "respx>=0.21",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["sentinel*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
